+++ CHANGELOG.md
@@ -11,4 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - plotBaseTiles function
 - Tile class with helper methods
 - Pan viewport helper
-- Coords and text drawing
+- Coords debug
+- Text drawing
+- Keyboard bindings